+++ R/genomics_by_sample.R
@@ -14,8 +14,8 @@
 #' This can be used in place of `sample_id`, `study_id`, `molecular_profile_id` arguments above if you
 #' need to pull samples from several different studies at once. If passed this will take overwrite `sample_id`, `study_id`, `molecular_profile_id` if also passed.
 #' @param data_type specify what type of data to return. Options are`mutations`, `cna`, `fusion`
-#' @param genes A vector of entrez ids. If NULL, it will return gene results for all available data for that sample.
-#' @param add_hugo Logical indicating whether `HugoSymbol` should be added to your results. cBioPortal API does not return this by default (only EntrezId) but this functions default is `TRUE` and adds this by default.
+#' @param genes A vector of entrez ids or Hugo symbols. If Hugo symbols are supplied, they will be converted to entrez ids using the `get_entrez_id()` function. If NULL, it will return gene results for all available data for that sample.
+#' @param add_hugo Logical indicating whether `HugoSymbol` should be added to your results. cBioPortal API does not return this by default (only EntrezId) but this function's default is `TRUE` and adds this by default.
 #' @param base_url The database URL to query
 #' If `NULL` will default to URL set with `set_cbioportal_db(<your_db>)`
 #'
@@ -91,7 +91,18 @@
     "fusion" = "fusion",
     "cna" = "discrete-copy-number")
 
-  resolved_genes <- genes
+  # check if hugo symbol or entrez id was supplied
+  genes_class <- class(genes)
+
+  # if character, convert to entrez ID
+
+  resolved_genes <- genes %>%
+    purrr::when(
+      is.character(.) ~ {
+        cli::cli_inform("Hugo symbols were converted to entrez IDs in order to query the cBioPortal API (see {.code ?get_entrez_id} for more info)")
+        get_entrez_id(., base_url = base_url)$entrezGeneId
+    },
+    TRUE ~ .)
 
   # Make Informed guesses on parameters -------------------------------------